meta: Initial commit
This commit adds a simple Discord bot that creates and responds to
select menus using Twilight.

The main purpose of this bot is to demonstrate the API introduced by
GH-PR twilight-rs/twilight#2219. Specifically, the code in this
repository is a "quick and dirty" test, and its code is not scalable,
idiomatic, or in any way recommendable outside testing the API.