main: Move token into separate block
This code moves the code that has access to the Discord bot token into
a separate code block. While this *doesn't* improve the security, it
should make it more obvious that the code doesn't do anything fishy with
the token.

Again, this is something only relevant for this example. This pattern
should not be applied in "real" bots.